[FIX/#180] AUTH Mapper 파일 구현
