+++ data/src/main/java/com/terning/data/dto/request/SignInRequestDto.kt
@@ -1,6 +1,5 @@
 package com.terning.data.dto.request
 
-import com.terning.domain.entity.request.SignInRequestModel
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -9,6 +8,3 @@ data class SignInRequestDto(
     @SerialName("authType")
     val authType: String
 )
-
-fun SignInRequestModel.toSignInRequestDto(): SignInRequestDto =
-    SignInRequestDto(authType = authType)
