+++ data/src/main/java/com/terning/data/mapper/auth/SignInRequestMapper.kt
@@ -1,0 +1,7 @@
+package com.terning.data.mapper.auth
+
+import com.terning.data.dto.request.SignInRequestDto
+import com.terning.domain.entity.auth.SignInRequestModel
+
+fun SignInRequestModel.toSignInRequestDto(): SignInRequestDto =
+    SignInRequestDto(authType = authType)
