+++ data/src/main/java/com/terning/data/mapper/auth/SignInResponseMapper.kt
@@ -1,0 +1,13 @@
+package com.terning.data.mapper.auth
+
+import com.terning.data.dto.response.SignInResponseDto
+import com.terning.domain.entity.auth.SignInResponseModel
+
+fun SignInResponseDto.toSignInModel(): SignInResponseModel =
+    SignInResponseModel(
+        accessToken = accessToken,
+        refreshToken = refreshToken,
+        userId = userId,
+        authId = authId,
+        authType = authType
+    )
