+++ domain/src/main/java/com/terning/domain/entity/auth/SignInResponseModel.kt
@@ -1,4 +1,4 @@
-package com.terning.domain.entity.response
+package com.terning.domain.entity.auth
 
 data class SignInResponseModel(
     val accessToken: String?,
