+++ domain/src/main/java/com/terning/domain/entity/auth/SignUpRequestModel.kt
@@ -1,4 +1,4 @@
-package com.terning.domain.entity.request
+package com.terning.domain.entity.auth
 
 data class SignUpRequestModel (
     val name : String,
