+++ domain/src/main/java/com/terning/domain/entity/auth/SignUpResponseModel.kt
@@ -1,4 +1,4 @@
-package com.terning.domain.entity.response
+package com.terning.domain.entity.auth
 
 data class SignUpResponseModel(
     val accessToken: String,
